Make unit test structed
