+++ test/test_library/library_schema/base_cfm.yaml
@@ -1,0 +1,7 @@
+---
+define:
+  name: Custom Flow Model
+  type: lmio/schema
+fields:
+  flow.id:
+    type: str
