+++ test/test_library/library_schema/base_no_fields.yaml
@@ -1,0 +1,3 @@
+---
+define:
+  type: lmio/schema
