+++ test/test_library/library_schema/document_list.yaml
@@ -1,0 +1,1 @@
+[]
