+++ test/test_library/library_schema/malformed.yaml
@@ -1,0 +1,1 @@
+fields: [
